+++ rye/src/cli/rye.rs
@@ -1,13 +1,22 @@
 use std::process::Command;
 
 use anyhow::{bail, Context, Error};
-use clap::Parser;
+use clap::{CommandFactory, Parser};
+use clap_complete::Shell;
 
 /// Rye self management
 #[derive(Parser, Debug)]
 pub struct Args {
     #[command(subcommand)]
     command: SubCommand,
+}
+
+/// Generates a completion script for a shell.
+#[derive(Parser, Debug)]
+pub struct CompletionCommand {
+    /// The shell to generate a completion script for (defaults to 'bash').
+    #[arg(short, long)]
+    shell: Option<Shell>,
 }
 
 /// Performs an update of rye.
@@ -29,13 +38,26 @@ pub struct UpdateCommand {
 
 #[derive(Parser, Debug)]
 enum SubCommand {
+    Completion(CompletionCommand),
     Update(UpdateCommand),
 }
 
 pub fn execute(cmd: Args) -> Result<(), Error> {
     match cmd.command {
+        SubCommand::Completion(args) => completion(args),
         SubCommand::Update(args) => update(args),
     }
+}
+
+fn completion(args: CompletionCommand) -> Result<(), Error> {
+    clap_complete::generate(
+        args.shell.unwrap_or(Shell::Bash),
+        &mut super::Args::command(),
+        "rye",
+        &mut std::io::stdout(),
+    );
+
+    Ok(())
 }
 
 fn update(args: UpdateCommand) -> Result<(), Error> {